Rc<RefCell<Box<...>>> -> Box<...> for exact solutions
This replaces runtime reference counting and borrow checking with
regular static borrow checking. We now have to unbox (and unwrap the
containing Option) every iteration of a for in order to bring the
immutable borrow into the same scope as the mutable borrow, relying on
(presumably) NLL to let us do this. Might be possible to improve
further.