Rework timestepping (WIP)
Introduce traits for timesteppers and for problems that can be
timestepped. Currently performs a lot worse than the previous in-problem
timestepper since it allocates a new RHS buffer each timestep, rather
than using a preallocated buffer as before. Restoring the buffer is WIP.
Includes a couple of tiny fixes (off-by-one errors in loops - need to
replace with proper iterators anyway...)