Kernel density esimation
A few simple kernels are also implemented. The values and indefinite
integrals of kernels are provided so that point densities and averages
over arbitrary regions can easily be calculated by `DensityEstimator`.
This is useful when using KDE for initial conditions for the cell
density as we actually store the (computational) cell-average density.