Use exact steady solution for phi_i as its IC
Actually, it's the steady solution for phi_i ignoring the cell cap part
of the boundary conditions. This makes it possible to find an exact
solution (since phi_i then decouples) and it is very close to the true
solution anyway - it's still a massive improvement over randomly
starting with uniform 0.9 or whatever.